+++ src/eosio.wps.cpp
@@ -1,10 +1,11 @@
 #include <eosio.wps/eosio.wps.hpp>
 
-#include "proposal.cpp"
-#include "vote.cpp"
-#include "on_notify.cpp"
 #include "clean.cpp"
-#include "periods.cpp"
-
-#include "settings.cpp"
+#include "drafts.cpp"
+#include "on_notify.cpp"
+#include "proposals.cpp"
 #include "state.cpp"
+#include "deposits.cpp"
+#include "periods.cpp"
+#include "settings.cpp"
+#include "vote.cpp"